t1

+++ src/main/java/net/azib/ipscan/core/ScanningProgressCallback.java
@@ -9,20 +9,20 @@ import java.net.InetAddress;
 
 /**
  * This callback is called on scanning state updates.
- * 
+ *
  * @author Anton Keks
  */
 public interface ScanningProgressCallback {
-			
+
 	/**
 	 * This method is called on scanning progress updates.
 	 * There are no guarantees that this method is called on every
-	 * scanning iteration. 
-	 * 
+	 * scanning iteration.
+	 *
 	 * @param currentAddress currently scanned IP address, can be null
 	 * @param runningThreads number of currently running threads
 	 * @param percentageComplete value from 0 to 100, showing how much work
 	 * 		is already done.
 	 */
-	void updateProgress(InetAddress currentAddress, int runningThreads, int percentageComplete);
+	void updateProgress(InetAddress currentAddress, int runningThreads, double percentageComplete);
 }
